# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: Team Task Manager API specification
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /users/me:
    get:
      operationId: getCurrentUser
      tags: [users]
      summary: Get current user profile (returns 404 if not registered yet)
      security:
        - bearerAuth: []
      responses:
        "200":
          description: Current user
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
        "401":
          description: Unauthorized
        "404":
          description: User not found - proceed to choose-role
    post:
      operationId: createCurrentUser
      tags: [users]
      summary: Create user record after initial signup (choose-role flow)
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateUserBody"
      responses:
        "201":
          description: User created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
        "200":
          description: User already exists
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
  /users/me/role:
    put:
      operationId: updateUserRole
      tags: [users]
      summary: Update a user's role (admin only)
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateUserRoleBody"
      responses:
        "200":
          description: Role updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
        "401":
          description: Unauthorized
  /users/me/profile:
    put:
      operationId: updateUserProfile
      tags: [users]
      summary: Update current user profile (name)
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateProfileBody"
      responses:
        "200":
          description: Profile updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
        "401":
          description: Unauthorized
  /users:
    get:
      operationId: listUsers
      tags: [users]
      summary: List all users (for member assignment)
      security:
        - bearerAuth: []
      responses:
        "200":
          description: List of users
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/User"
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
  schemas:
...
    User:
      type: object
      properties:
        id:
          type: string
        clerkId:
          type: string
        name:
          type: string
        email:
          type: string
        role:
          type: string
          enum: [admin, member]
        createdAt:
          type: string
          format: date-time
      required:
        - id
        - clerkId
        - name
        - email
        - role
        - createdAt
    CreateUserBody:
      type: object
      properties:
        role:
          type: string
          enum: [admin, member]
      required:
        - role
    UpdateUserRoleBody:
      type: object
      properties:
        role:
          type: string
          enum: [admin, member]
        targetUserId:
          type: string
      required:
        - role
        - targetUserId
    UpdateProfileBody:
      type: object
      properties:
        name:
          type: string
      required:
        - name
...