# lib/api-zod/src/generated/types/updateUserRoleBody.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Team Task Manager API specification
 * OpenAPI spec version: 0.1.0
 */
import type { UpdateUserRoleBodyRole } from "./updateUserRoleBodyRole";

export interface UpdateUserRoleBody {
  role: UpdateUserRoleBodyRole;
  targetUserId: string;
}